agents/template-agent/pyproject.toml
[project]
name = "template-agent"
version = "0.1"
description = "template-agent agent"
readme = "README.md"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.8",
    "fuzzywuzzy>=0.18.0",
    "kagent-langgraph>=0.7.13",
    "kubernetes>=35.0.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.9",
    "langdetect>=1.0.9",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-cli>=0.4.12",
    "langgraph-sdk>=0.3.5",
    "openai>=2.20.0",
    "psycopg[binary,pool]>=3.3.2",
    "python-levenshtein>=0.27.3",
]
